Support Composite SLOs
Composite SLOs were added into the spec in OpenSLO/OpenSLO#182, however they never made it into the Go struct code.

Essentially, with the new spec, Indicators can be added in the Objective field, as a inline, or ref. A new field is also added for the weight.

By allowing this for Composite SLOs we need to remove the required validation for Indicator within the main SLO Spec.